+++ src/storage-mover/azext_storage_mover/_params.py
@@ -58,3 +58,15 @@ def load_arguments(self, _):  # pylint: disable=unused-argument
         with self.argument_context('storage-mover endpoint ' + command) as c:
             c.argument('password_uri', help='The Azure Key Vault secret URI which stores the password. Use empty string to clean-up existing value.')
             c.argument('username_uri', help='The Azure Key Vault secret URI which stores the username. Use empty string to clean-up existing value.')
+
+    with self.argument_context('storage-mover endpoint create-for-s3-with-hmac') as c:
+        c.argument('source_uri', help='The URI which points to the S3-compatible source.')
+        c.argument('source_type', arg_type=get_enum_type(['BACKBLAZE', 'CLOUDFLARE', 'GCS', 'IBM', 'MINIO']),
+                   help='The source type of the S3WithHmac endpoint.')
+        c.argument('other_source_type_description',
+                   help='The description for other source type of S3WithHmac endpoint.')
+
+    for command in ['create-for-s3-with-hmac', 'update-for-s3-with-hmac']:
+        with self.argument_context('storage-mover endpoint ' + command) as c:
+            c.argument('access_key_uri', help='The Azure Key Vault secret URI which stores the access key. Use empty string to clean-up existing value.')
+            c.argument('secret_key_uri', help='The Azure Key Vault secret URI which stores the secret key. Use empty string to clean-up existing value.')