+++ src/storage-mover/azext_storage_mover/aaz/latest/storage_mover/endpoint/_list.py
@@ -15,16 +15,13 @@ from azure.cli.core.aaz import *
     "storage-mover endpoint list",
 )
 class List(AAZCommand):
-    """Lists all Endpoints in a Storage Mover.
-
-    :example: endpoint list
-        az storage-mover endpoint list -g {rg} --storage-mover-name {mover_name}
+    """List all Endpoints in a Storage Mover.
     """
 
     _aaz_info = {
-        "version": "2025-07-01",
+        "version": "2025-12-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/endpoints", "2025-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/endpoints", "2025-12-01"],
         ]
     }
 
@@ -52,6 +49,9 @@ class List(AAZCommand):
             options=["--storage-mover-name"],
             help="The name of the Storage Mover resource.",
             required=True,
+            fmt=AAZStrArgFormat(
+                pattern="^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
+            ),
         )
         return cls._args_schema
 
@@ -121,7 +121,7 @@ class List(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-07-01",
+                    "api-version", "2025-12-01",
                     required=True,
                 ),
             }
@@ -214,6 +214,9 @@ class List(AAZCommand):
 
             properties = cls._schema_on_200.value.Element.properties
             properties.description = AAZStrType()
+            properties.endpoint_kind = AAZStrType(
+                serialized_name="endpointKind",
+            )
             properties.endpoint_type = AAZStrType(
                 serialized_name="endpointType",
                 flags={"required": True},
@@ -272,6 +275,29 @@ class List(AAZCommand):
             )
             disc_nfs_mount.nfs_version = AAZStrType(
                 serialized_name="nfsVersion",
+            )
+
+            disc_s3_with_hmac = cls._schema_on_200.value.Element.properties.discriminate_by("endpoint_type", "S3WithHMAC")
+            disc_s3_with_hmac.credentials = AAZObjectType()
+            disc_s3_with_hmac.other_source_type_description = AAZStrType(
+                serialized_name="otherSourceTypeDescription",
+            )
+            disc_s3_with_hmac.source_type = AAZStrType(
+                serialized_name="sourceType",
+            )
+            disc_s3_with_hmac.source_uri = AAZStrType(
+                serialized_name="sourceUri",
+            )
+
+            credentials = cls._schema_on_200.value.Element.properties.discriminate_by("endpoint_type", "S3WithHMAC").credentials
+            credentials.access_key_uri = AAZStrType(
+                serialized_name="accessKeyUri",
+            )
+            credentials.secret_key_uri = AAZStrType(
+                serialized_name="secretKeyUri",
+            )
+            credentials.type = AAZStrType(
+                flags={"required": True},
             )
 
             disc_smb_mount = cls._schema_on_200.value.Element.properties.discriminate_by("endpoint_type", "SmbMount")