+++ src/storage-mover/azext_storage_mover/aaz/latest/storage_mover/endpoint/_update.py
@@ -12,13 +12,13 @@ from azure.cli.core.aaz import *
 
 
 class Update(AAZCommand):
-    """Updates an Endpoint resource, which represents a data transfer source or destination.
+    """Update an Endpoint resource, which represents a data transfer source or destination.
     """
 
     _aaz_info = {
-        "version": "2025-07-01",
+        "version": "2025-12-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/endpoints/{}", "2025-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/endpoints/{}", "2025-12-01"],
         ]
     }
 
@@ -54,6 +54,9 @@ class Update(AAZCommand):
             help="The name of the Storage Mover resource.",
             required=True,
             id_part="name",
+            fmt=AAZStrArgFormat(
+                pattern="^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
+            ),
         )
 
         # define Arg Group "Identity"
@@ -61,12 +64,6 @@ class Update(AAZCommand):
         # define Arg Group "Properties"
 
         _args_schema = cls._args_schema
-        _args_schema.storage_blob_container = AAZObjectArg(
-            options=["--storage-blob-container"],
-            arg_group="Properties",
-            help="Storage Blob Container Object",
-        )
-
         _args_schema.azure_multi_cloud_connector = AAZObjectArg(
             options=["--azure-multi-cloud-connector"],
             arg_group="Properties",
@@ -80,12 +77,6 @@ class Update(AAZCommand):
             arg_group="Properties",
             help="A description for the Endpoint.",
             nullable=True,
-        )
-
-        storage_blob_container = cls._args_schema.storage_blob_container
-        storage_blob_container.storage_account_resource_id = AAZResourceIdArg(
-            options=["storage-account-resource-id"],
-            help="The Azure Resource ID of the storage account that is the target destination.",
         )
 
         azure_multi_cloud_connector = cls._args_schema.azure_multi_cloud_connector
@@ -196,7 +187,7 @@ class Update(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-07-01",
+                    "api-version", "2025-12-01",
                     required=True,
                 ),
             }
@@ -283,7 +274,7 @@ class Update(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-07-01",
+                    "api-version", "2025-12-01",
                     required=True,
                 ),
             }
@@ -347,13 +338,8 @@ class Update(AAZCommand):
             properties = _builder.get(".properties")
             if properties is not None:
                 properties.set_prop("description", AAZStrType, ".description")
-                properties.discriminate_by("endpointType", "AzureStorageBlobContainer")
                 properties.discriminate_by("endpointType", "AzureMultiCloudConnector")
                 properties.discriminate_by("endpointType", "SmbMount")
-
-            disc_azure_storage_blob_container = _builder.get(".properties{endpointType:AzureStorageBlobContainer}")
-            if disc_azure_storage_blob_container is not None:
-                disc_azure_storage_blob_container.set_prop("storageAccountResourceId", AAZStrType, ".storage_blob_container.storage_account_resource_id", typ_kwargs={"flags": {"required": True}})
 
             disc_azure_multi_cloud_connector = _builder.get(".properties{endpointType:AzureMultiCloudConnector}")
             if disc_azure_multi_cloud_connector is not None:
@@ -366,7 +352,6 @@ class Update(AAZCommand):
             credentials = _builder.get(".properties{endpointType:SmbMount}.credentials")
             if credentials is not None:
                 credentials.set_prop("passwordUri", AAZStrType, ".password_uri")
-                credentials.set_const("type", "AzureKeyVaultSmb", AAZStrType, ".", typ_kwargs={"flags": {"required": True}})
                 credentials.set_prop("usernameUri", AAZStrType, ".username_uri")
 
             return _instance_value
@@ -448,6 +433,9 @@ class _UpdateHelper:
 
         properties = _schema_endpoint_read.properties
         properties.description = AAZStrType()
+        properties.endpoint_kind = AAZStrType(
+            serialized_name="endpointKind",
+        )
         properties.endpoint_type = AAZStrType(
             serialized_name="endpointType",
             flags={"required": True},
@@ -506,6 +494,29 @@ class _UpdateHelper:
         )
         disc_nfs_mount.nfs_version = AAZStrType(
             serialized_name="nfsVersion",
+        )
+
+        disc_s3_with_hmac = _schema_endpoint_read.properties.discriminate_by("endpoint_type", "S3WithHMAC")
+        disc_s3_with_hmac.credentials = AAZObjectType()
+        disc_s3_with_hmac.other_source_type_description = AAZStrType(
+            serialized_name="otherSourceTypeDescription",
+        )
+        disc_s3_with_hmac.source_type = AAZStrType(
+            serialized_name="sourceType",
+        )
+        disc_s3_with_hmac.source_uri = AAZStrType(
+            serialized_name="sourceUri",
+        )
+
+        credentials = _schema_endpoint_read.properties.discriminate_by("endpoint_type", "S3WithHMAC").credentials
+        credentials.access_key_uri = AAZStrType(
+            serialized_name="accessKeyUri",
+        )
+        credentials.secret_key_uri = AAZStrType(
+            serialized_name="secretKeyUri",
+        )
+        credentials.type = AAZStrType(
+            flags={"required": True},
         )
 
         disc_smb_mount = _schema_endpoint_read.properties.discriminate_by("endpoint_type", "SmbMount")