+++ src/storage-mover/azext_storage_mover/aaz/latest/storage_mover/endpoint/_wait.py
@@ -20,7 +20,7 @@ class Wait(AAZWaitCommand):
 
     _aaz_info = {
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/endpoints/{}", "2025-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/endpoints/{}", "2025-12-01"],
         ]
     }
 
@@ -54,6 +54,9 @@ class Wait(AAZWaitCommand):
             help="The name of the Storage Mover resource.",
             required=True,
             id_part="name",
+            fmt=AAZStrArgFormat(
+                pattern="^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
+            ),
         )
         return cls._args_schema
 
@@ -126,7 +129,7 @@ class Wait(AAZWaitCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-07-01",
+                    "api-version", "2025-12-01",
                     required=True,
                 ),
             }
@@ -208,6 +211,9 @@ class Wait(AAZWaitCommand):
 
             properties = cls._schema_on_200.properties
             properties.description = AAZStrType()
+            properties.endpoint_kind = AAZStrType(
+                serialized_name="endpointKind",
+            )
             properties.endpoint_type = AAZStrType(
                 serialized_name="endpointType",
                 flags={"required": True},
@@ -266,6 +272,29 @@ class Wait(AAZWaitCommand):
             )
             disc_nfs_mount.nfs_version = AAZStrType(
                 serialized_name="nfsVersion",
+            )
+
+            disc_s3_with_hmac = cls._schema_on_200.properties.discriminate_by("endpoint_type", "S3WithHMAC")
+            disc_s3_with_hmac.credentials = AAZObjectType()
+            disc_s3_with_hmac.other_source_type_description = AAZStrType(
+                serialized_name="otherSourceTypeDescription",
+            )
+            disc_s3_with_hmac.source_type = AAZStrType(
+                serialized_name="sourceType",
+            )
+            disc_s3_with_hmac.source_uri = AAZStrType(
+                serialized_name="sourceUri",
+            )
+
+            credentials = cls._schema_on_200.properties.discriminate_by("endpoint_type", "S3WithHMAC").credentials
+            credentials.access_key_uri = AAZStrType(
+                serialized_name="accessKeyUri",
+            )
+            credentials.secret_key_uri = AAZStrType(
+                serialized_name="secretKeyUri",
+            )
+            credentials.type = AAZStrType(
+                flags={"required": True},
             )
 
             disc_smb_mount = cls._schema_on_200.properties.discriminate_by("endpoint_type", "SmbMount")