+++ src/storage-mover/azext_storage_mover/aaz/latest/storage_mover/job_definition/_wait.py
@@ -20,7 +20,7 @@ class Wait(AAZWaitCommand):
 
     _aaz_info = {
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/projects/{}/jobdefinitions/{}", "2025-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/projects/{}/jobdefinitions/{}", "2025-12-01"],
         ]
     }
 
@@ -60,6 +60,9 @@ class Wait(AAZWaitCommand):
             help="The name of the Storage Mover resource.",
             required=True,
             id_part="name",
+            fmt=AAZStrArgFormat(
+                pattern="^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
+            ),
         )
         return cls._args_schema
 
@@ -136,7 +139,7 @@ class Wait(AAZWaitCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-07-01",
+                    "api-version", "2025-12-01",
                     required=True,
                 ),
             }
@@ -194,9 +197,13 @@ class Wait(AAZWaitCommand):
                 serialized_name="agentResourceId",
                 flags={"read_only": True},
             )
+            properties.connections = AAZListType()
             properties.copy_mode = AAZStrType(
                 serialized_name="copyMode",
                 flags={"required": True},
+            )
+            properties.data_integrity_validation = AAZStrType(
+                serialized_name="dataIntegrityValidation",
             )
             properties.description = AAZStrType()
             properties.job_type = AAZStrType(
@@ -214,10 +221,14 @@ class Wait(AAZWaitCommand):
                 serialized_name="latestJobRunStatus",
                 flags={"read_only": True},
             )
+            properties.preserve_permissions = AAZBoolType(
+                serialized_name="preservePermissions",
+            )
             properties.provisioning_state = AAZStrType(
                 serialized_name="provisioningState",
                 flags={"read_only": True},
             )
+            properties.schedule = AAZObjectType()
             properties.source_name = AAZStrType(
                 serialized_name="sourceName",
                 flags={"required": True},
@@ -243,6 +254,48 @@ class Wait(AAZWaitCommand):
             properties.target_subpath = AAZStrType(
                 serialized_name="targetSubpath",
             )
+
+            connections = cls._schema_on_200.properties.connections
+            connections.Element = AAZStrType()
+
+            schedule = cls._schema_on_200.properties.schedule
+            schedule.cron_expression = AAZStrType(
+                serialized_name="cronExpression",
+            )
+            schedule.days_of_month = AAZListType(
+                serialized_name="daysOfMonth",
+            )
+            schedule.days_of_week = AAZListType(
+                serialized_name="daysOfWeek",
+            )
+            schedule.end_date = AAZStrType(
+                serialized_name="endDate",
+            )
+            schedule.execution_time = AAZObjectType(
+                serialized_name="executionTime",
+            )
+            schedule.frequency = AAZStrType(
+                flags={"required": True},
+            )
+            schedule.is_active = AAZBoolType(
+                serialized_name="isActive",
+                flags={"required": True},
+            )
+            schedule.start_date = AAZStrType(
+                serialized_name="startDate",
+            )
+
+            days_of_month = cls._schema_on_200.properties.schedule.days_of_month
+            days_of_month.Element = AAZIntType()
+
+            days_of_week = cls._schema_on_200.properties.schedule.days_of_week
+            days_of_week.Element = AAZStrType()
+
+            execution_time = cls._schema_on_200.properties.schedule.execution_time
+            execution_time.hour = AAZIntType(
+                flags={"required": True},
+            )
+            execution_time.minute = AAZFloatType()
 
             source_target_map = cls._schema_on_200.properties.source_target_map
             source_target_map.value = AAZListType(