+++ src/storage-mover/azext_storage_mover/aaz/latest/storage_mover/job_run/_show.py
@@ -15,16 +15,13 @@ from azure.cli.core.aaz import *
     "storage-mover job-run show",
 )
 class Show(AAZCommand):
-    """Gets a Job Run resource.
-
-    :example: job-run show
-        az storage-mover job-run show -n {job_name} -g {rg} --job-definition-name {job_definition} --project-name {project_name} --storage-mover-name {mover_name}
+    """Get a Job Run resource.
     """
 
     _aaz_info = {
-        "version": "2025-07-01",
+        "version": "2025-12-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/projects/{}/jobdefinitions/{}/jobruns/{}", "2025-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/projects/{}/jobdefinitions/{}/jobruns/{}", "2025-12-01"],
         ]
     }
 
@@ -70,6 +67,9 @@ class Show(AAZCommand):
             help="The name of the Storage Mover resource.",
             required=True,
             id_part="name",
+            fmt=AAZStrArgFormat(
+                pattern="^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
+            ),
         )
         return cls._args_schema
 
@@ -150,7 +150,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-07-01",
+                    "api-version", "2025-12-01",
                     required=True,
                 ),
             }
@@ -284,6 +284,10 @@ class Show(AAZCommand):
                 serialized_name="scanStatus",
                 flags={"read_only": True},
             )
+            properties.scheduled_execution_time = AAZStrType(
+                serialized_name="scheduledExecutionTime",
+                flags={"read_only": True},
+            )
             properties.source_name = AAZStrType(
                 serialized_name="sourceName",
                 flags={"read_only": True},
@@ -311,11 +315,26 @@ class Show(AAZCommand):
                 serialized_name="targetResourceId",
                 flags={"read_only": True},
             )
+            properties.trigger_type = AAZStrType(
+                serialized_name="triggerType",
+                flags={"read_only": True},
+            )
+            properties.warnings = AAZListType(
+                flags={"read_only": True},
+            )
 
             error = cls._schema_on_200.properties.error
             error.code = AAZStrType()
             error.message = AAZStrType()
             error.target = AAZStrType()
+
+            warnings = cls._schema_on_200.properties.warnings
+            warnings.Element = AAZObjectType()
+
+            _element = cls._schema_on_200.properties.warnings.Element
+            _element.code = AAZStrType()
+            _element.message = AAZStrType()
+            _element.target = AAZStrType()
 
             system_data = cls._schema_on_200.system_data
             system_data.created_at = AAZStrType(